Create #Java If…else Statement
The if statements are executed from the top towards a bottom. As soon as that test expression is true, the codes inside the body of that if statement is executed. Then, the control flow of the program jumps outside the if-else-if ladder. If all the test expressions are false, codes inside that body of else are executed.